add log prefix in non-dp mode engine core
